Update Google-API-Project
This project shows my experience of working with the Google API's. The program uses a Tkinter entry widget which allows the user to input a location that they would like to see on the google map and it brings that location onto the map. As well as it gives the user the option to view the map between roadmap, satellite, terrain, and hybrid map view. It also allows the user to zoom into their location.